Function zmq_rs::gen_curve_keypair

pub fn gen_curve_keypair() -> Result<(String, String), Error>

Generate z85-encoded public and private keypair with libsodium.

Binding of int zmq_curve_keypair (char *z85_public_key, char *z85_secret_key);

The function will return a newly generated random keypair consisting of a public key and a secret key. The keys are encoded using z85_encode().